backend/requirements.txt
flask==3.1.2
pydantic==2.12.5
pydantic-core==2.41.5
flask-cors==6.0.1
numpy==1.26.4
pytest==7.4.3
requests==2.32.5
flasgger==0.9.7.1
Flask-Limiter==3.5.1
python-dotenv==1.0.0
